+++ src/components/LoanFinding/PartnerSpotlightSection.vue
@@ -1,19 +1,31 @@
 <template>
-	<div v-if="mfiRecommendationsLoans.length > 0" class="tw-w-full tw-bg-secondary">
+	<div v-if="loans.length > 0" class="tw-w-full tw-bg-secondary">
 		<div class="tw-mx-auto tw-px-0 md:tw-px-4 lg:tw-px-8" style="max-width: 1200px;">
-			<m-f-i-hero />
+			<m-f-i-hero
+				section-title="IMPACT AT KIVA"
+				:headline="headline"
+				:subheadline="subheadline"
+				:partner-image="partnerImage"
+				:partner-image-footer="partnerImageFooter"
+				:partner-text="partnerText"
+				:subheads-title="subheadsTitle"
+				:subheads="subheads"
+			/>
 		</div>
 
 		<lending-category-section
-			title="Support Borrowers from Fundación Pro Mujer"
-			:loans="mfiRecommendationsLoans"
+			id="mfiCarousel"
+			class="tw-mt-5"
+			:title="carouselTitle"
+			:loans="loans"
 			:enable-loan-card-exp="enableLoanCardExp"
+			@add-to-basket="$emit('add-to-basket', $event)"
 		/>
 
 		<div class="tw-flex tw-justify-center tw-mt-4">
 			<a
 				class="tw-text-action tw-text-base"
-				href="https://www.kiva.org/lend/filter?partner=59"
+				:href="viewAllLink"
 				v-kv-track-event="['event-tracking', 'click', 'mfi-view-all']"
 			>
 				View All
@@ -25,7 +37,6 @@
 <script>
 import MFIHero from '@/components/LoansByCategory/MFIRecommendations/MFIHero';
 import LendingCategorySection from '@/components/LoanFinding/LendingCategorySection';
-import mfiRecommendationsLoans from '@/graphql/query/lendByCategory/mfiRecommendationsLoans.graphql';
 
 export default {
 	name: 'PartnerSpotlightSection',
@@ -38,32 +49,55 @@ export default {
 		enableLoanCardExp: {
 			type: Boolean,
 			default: false
-		}
-	},
-	data() {
-		return {
-			mfiRecommendationsLoans: [],
-		};
+		},
+		spotlightData: {
+			type: Object,
+			default: () => {}
+		},
+		loans: {
+			type: Array,
+			default: () => []
+		},
 	},
 	computed: {
-		selectedChannelLoanIds() {
-			return this.mfiRecommendationsLoans?.map(element => element.id) ?? [];
+		headline() {
+			return this.spotlightData?.headline ?? '';
 		},
-	},
-	methods: {
-		fetchMFILoans() {
-			// Load mfi recommendations loans data
-			return this.apollo.query({
-				query: mfiRecommendationsLoans,
-			}).then(({ data }) => {
-				this.mfiRecommendationsLoans = data?.fundraisingLoans?.values ?? [];
-				const numberLoans = this.mfiRecommendationsLoans.length;
-				if (!numberLoans) this.$kvTrackEvent('event-tracking', 'update', 'mfi-no-featured-loan-available');
-			});
+		subheadline() {
+			return this.spotlightData?.subheadline ?? '';
 		},
-	},
-	mounted() {
-		this.fetchMFILoans();
+		partnerImage() {
+			return this.spotlightData?.image ?? '';
+		},
+		partnerImageFooter() {
+			return this.spotlightData?.imageFooter ?? '';
+		},
+		partnerText() {
+			return this.spotlightData?.partnerText ?? '';
+		},
+		carouselTitle() {
+			return this.spotlightData?.carouselTitle ?? '';
+		},
+		subheadsTitle() {
+			return this.spotlightData?.subheadsTitle ?? '';
+		},
+		subheads() {
+			return this.spotlightData?.subheads ?? [];
+		},
+		viewAllLink() {
+			return this.spotlightData?.viewAllLink ?? '';
+		}
 	},
 };
 </script>
+
+<style lang="postcss" scoped>
+
+#mfiCarousel >>> h2 {
+	@apply tw-text-h4 tw-mb-0 tw-text-action tw-ml-2;
+}
+
+#mfiCarousel >>> section {
+	@apply tw-mt-1;
+}
+</style>